JavaScript form check
Added some JavaScript to make sure that the user inputs a file (Over
write the standard Chrome file upload check due to having to check the
select input dialog) as well as making sure that the user selected a
problem number, and does not try to submit the "Please choose a problem
number" option. This has been hidden and disabled so it will not appear
in the selection dialog.