SQL, PHP, scoreboard fixes
Fixed some SQL statements that retrieve data.
Redid some of Andriy's scoreboard code as there were a few bugs. Nothing
big, simple a missing <td></td>, and a line of code that reset the
total_time to 0 every iteration of the problems.
Created a new MySQL table named "data" that has two columns, "name" and
"value". Inside that table is a single entry (as of now) named
"start_time" with a value, in seconds, of when the event started.
Added some math statements that calculated, correctly, what the score is
for the teams.

What's left to do:
Show the order of the teams as ranks, instead of just based off how that
MySQL table orders them (Which is unranked).