Formatted scoreboard design
Made the "content" div 15% bigger (80% -> 95%)
Renamed "Trys" to "Attempts" in the table header
Converted $complete to a string
Instead of displaying how many problems the team has completed, the
scoreboard now shows which problems each team completed. This is
represented as "1, 2, 3, etc." for the completed problems.

Note: One of the last few lines of PHP removes the last 2 characters off
of the $complete string. This is because the last 2 characters are ", "
which make the resulting string appear incorrectly.