Scoreboard navigation
The LCPS HSPC navigation button will now take you to the scoreboard. On
the scoreboard page the same navigation button will take you to the
submissions page. This is most likely not the permanent solution but
just a temporary solution to make it easy to move around the pages
without manually entering their links.