Got class?
Needed a way to rank the teams based on their score. Created a super
simple PHP class that stores a teams total_time and then html_data which
is a formatted table of all the data for that team to be printed out.

The TeamData objects are then stored into an array, later looped over,
and this is where the teams need to be sorted based on their total_time
in ascending order (Smallest to greatest, so winner has the smallest
score, we know this). Should be fairly simple to do this.

Formatted some more things, removed some useless comments.