Team ordering
The table row for each team is now stored into a teamData class which
also has the total_time for that team.
Used a usort to sort the array of teamData's in ascending order (Lowest
to greatest score).
Foreach loop to display each teams table row.

Need to rewrite how the code gets the team_name. Should NOT be using
$_COOKIE["n"]